+++ pkg/network/errors.go
@@ -1,7 +1,0 @@
-// Package network provides support tools for working with network devices.
-package network
-
-import "errors"
-
-// ErrNotSet indicates that DNS can't be set.
-var ErrNotSet = errors.New("DNS can't be set")